+++ azure_devops_rust_api/src/pipelines/models.rs
@@ -279,15 +279,16 @@ pub mod pipeline_configuration {
         #[serde(rename = "designerHyphenJson")]
         DesignerHyphenJson,
     }
-    #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+    #[derive(Clone, Debug, PartialEq, Serialize, Deserialize, Default)]
     pub struct Repository {
-        pub id: String,
-        #[serde(rename = "type")]
-        pub type_: String,
+        #[serde(default, skip_serializing_if = "Option::is_none")]
+        pub id: Option<String>,
+        #[serde(rename = "type", default, skip_serializing_if = "Option::is_none")]
+        pub type_: Option<String>,
     }
     impl Repository {
-        pub fn new(id: String, type_: String) -> Self {
-            Self { id, type_ }
+        pub fn new() -> Self {
+            Self::default()
         }
     }
 }
@@ -435,14 +436,19 @@ pub struct Run {
     pub created_date: time::OffsetDateTime,
     #[serde(rename = "finalYaml", default, skip_serializing_if = "Option::is_none")]
     pub final_yaml: Option<String>,
-    #[serde(rename = "finishedDate", with = "crate::date_time::rfc3339")]
-    pub finished_date: time::OffsetDateTime,
+    #[serde(
+        rename = "finishedDate",
+        default,
+        with = "crate::date_time::rfc3339::option"
+    )]
+    pub finished_date: Option<time::OffsetDateTime>,
     #[doc = "A reference to a Pipeline."]
     pub pipeline: PipelineReference,
     #[doc = ""]
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub resources: Option<RunResources>,
-    pub result: run::Result,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub result: Option<run::Result>,
     pub state: run::State,
     pub url: String,
     #[serde(default, skip_serializing_if = "Option::is_none")]
@@ -453,9 +459,7 @@ impl Run {
         run_reference: RunReference,
         links: ReferenceLinks,
         created_date: time::OffsetDateTime,
-        finished_date: time::OffsetDateTime,
         pipeline: PipelineReference,
-        result: run::Result,
         state: run::State,
         url: String,
     ) -> Self {
@@ -464,10 +468,10 @@ impl Run {
             links,
             created_date,
             final_yaml: None,
-            finished_date,
+            finished_date: None,
             pipeline,
             resources: None,
-            result,
+            result: None,
             state,
             url,
             variables: None,