+++ vsts-api-patcher/src/main.rs
@@ -365,7 +365,6 @@ impl Patcher {
                                 "type": "string"
                             }
                         },
-                        "required": ["id", "type"]
                     }
                 })
             }
@@ -1154,12 +1153,13 @@ impl Patcher {
             (
                 "pipelines.json",
                 "Run",
+                // Excluded
+                // - finishedDate
+                // - result
                 r#"[
                     "_links",
                     "createdDate",
-                    "finishedDate",
                     "pipeline",
-                    "result",
                     "state",
                     "url"
                 ]"#,