+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed definition of git `PolicyConfiguration`
+- Fixed pipeline `Run` required fields
+  - Changed `finishedDate` and `result` to optional, as they are not present if run is in progress
+- Change pipeline `Repository` fields to be optional: `id`, `type`
 
 ### Added